refactor(github_review): extract ledger schema to _ledger_schema.py
Move _SCHEMA_VERSION/_DIRECTORY_MODE/_DATABASE_MODE/_SCHEMA + dataclasses
ReviewOperationRecord/ReviewAttemptRecord/MutationSlot from ledger.py
(868 → 772 lines) into _ledger_schema.py. ledger.py imports them and
re-exports for sibling modules; existing test imports keep working.

Refs: #4664
